pyproject.toml
[project]
name = "llama-server-home"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "loguru>=0.7.3",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "pynvml>=13.0.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit"
]


[tool.ruff]
line-length = 120
lint.select = ["E", "F", "W", "I", "N"]

[project.scripts]
list-nodes = "lsh.controller.cmd:list_nodes"
run-node = "lsh.node.cmd:main"
